+++ src/components/pages/DemoPage.tsx
@@ -3,8 +3,6 @@ import styled from "styled-components";
 import scrollToEvent from "../../module/common/scrollToEvent";
 import DashboardSection from "../sections/dashboard/DashboardSection";
 import DashboardSideMenu from "../sections/dashboard/DashboardSideMenu";
-import { useSelector } from "react-redux";
-import { AnalyzedMessage } from "../../@types/index.d";
 
 const DashboardBox = styled.div`
   margin-top: 80px;
@@ -26,27 +24,14 @@ const DashboardBox = styled.div`
 `;
 
 const DashboardPage = () => {
-  const analyzedMessages = useSelector(
-    (state: { analyzedMessagesSlice: AnalyzedMessage[] }) => state.analyzedMessagesSlice
-  );
-  const selectedChatRoomIndex = useSelector(
-    (state: { selectedRoomIndexSlice: number }) => state.selectedRoomIndexSlice
-  );
-
   useEffect(() => {
     scrollToEvent(0, "auto");
   }, []);
 
   return (
     <DashboardBox>
-      <DashboardSideMenu
-        analyzedMessages={analyzedMessages}
-        selectedChatRoomIndex={selectedChatRoomIndex}
-      />
-      <DashboardSection
-        analyzedMessages={analyzedMessages}
-        selectedChatRoomIndex={selectedChatRoomIndex}
-      />
+      {/* <DashboardSideMenu />
+      <DashboardSection /> */}
     </DashboardBox>
   );
 };
